Work around issue in Truffle DSL
- nested classes do not generate code any more, might be fixed later?
- classes need specializations to generated code, specialization in abstract super class is not sufficient
