+++ src/som/interpreter/nodes/specialized/AndMessageNode.java
@@ -4,22 +4,18 @@ import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.dsl.GenerateNodeFactory;
 import com.oracle.truffle.api.dsl.NodeFactory;
 import com.oracle.truffle.api.dsl.Specialization;
-import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.instrumentation.Tag;
 import com.oracle.truffle.api.nodes.DirectCallNode;
 import com.oracle.truffle.api.source.SourceSection;
 
 import bd.primitives.Primitive;
 import bd.primitives.Specializer;
-import bd.tools.nodes.Operation;
 import som.VM;
 import som.interpreter.nodes.ExpressionNode;
 import som.interpreter.nodes.literals.BlockNode;
-import som.interpreter.nodes.nary.BinaryBasicOperation;
 import som.interpreter.nodes.nary.BinaryComplexOperation;
 import som.interpreter.nodes.nary.BinaryExpressionNode;
 import som.interpreter.nodes.specialized.AndMessageNode.AndOrSplzr;
-import som.interpreter.nodes.specialized.AndMessageNodeFactory.AndBoolMessageNodeFactory;
 import som.vmobjects.SBlock;
 import som.vmobjects.SInvokable;
 import som.vmobjects.SSymbol;
@@ -104,35 +100,6 @@ public abstract class AndMessageNode extends BinaryComplexOperation {
       return false;
     } else {
       return (boolean) blockValueSend.call(new Object[] {argument});
-    }
-  }
-
-  @GenerateNodeFactory
-  public abstract static class AndBoolMessageNode extends BinaryBasicOperation
-      implements Operation {
-    @Override
-    protected boolean hasTagIgnoringEagerness(final Class<? extends Tag> tag) {
-      if (tag == OpComparison.class) {
-        return true;
-      } else {
-        return super.hasTagIgnoringEagerness(tag);
-      }
-    }
-
-    @Specialization
-    public final boolean doAnd(final VirtualFrame frame, final boolean receiver,
-        final boolean argument) {
-      return receiver && argument;
-    }
-
-    @Override
-    public String getOperation() {
-      return "&&";
-    }
-
-    @Override
-    public int getNumArguments() {
-      return 2;
     }
   }
 }
